refactor(layout): split CategoryPage into reusable PageLayout and smart wrapper
