--- src/shared/components/CategoryPage/components/CategoryHeader/index.ts
@@ -1,1 +1,0 @@
-export * from './CategoryHeader';
