+++ src/shared/components/PageHeader/index.ts
@@ -1,0 +1,1 @@
+export * from './PageHeader';
