+++ src/shared/components/PageLayout/PageLayout.tsx
@@ -1,0 +1,21 @@
+import styles from './PageLayout.module.scss';
+
+interface Props {
+  topNav: React.ReactNode;
+  header: React.ReactNode;
+  children: React.ReactNode;
+}
+
+export const PageLayout = ({ topNav, header, children }: Props) => {
+  return (
+    <main className={styles.pageLayout}>
+      {topNav}
+
+      <section className={styles.content}>
+        {header}
+
+        {children}
+      </section>
+    </main>
+  );
+};
