+++ src/shared/components/ProductsList/hooks/useProductsPagination.ts
