+++ src/shared/components/ProductsList/index.ts
